+++ frontend/src/components/Button/Button.jsx
@@ -12,5 +12,3 @@ export const Button = ({ children, ...props }) =>
 Button.propTypes = {
   children: PropTypes.node.isRequired
 }
-
-export default Button
