+++ frontend/src/components/PasswordValidator/index.js
@@ -1,0 +1,1 @@
+export * from './PasswordValidator'
