+++ frontend/src/components/index.js
@@ -1,0 +1,4 @@
+export * from './Button'
+export * from './Form'
+export * from './TextField'
+export * from './PasswordValidator'
