+++ frontend/src/index.js
@@ -1,7 +1,7 @@
 import "./index.css"
-import AppRouter from "./routes"
 import React from "react"
 import ReactDOM from "react-dom/client"
+import { AppRouter } from "."
 
 const root = ReactDOM.createRoot(document.getElementById("root"))
 root.render(
@@ -9,3 +9,8 @@ root.render(
     <AppRouter />
   </React.StrictMode>
 )
+
+export * from './components'
+export * from './pages'
+export * from './assets'
+export * from './routes'
