+++ frontend/src/pages/index.js
@@ -1,0 +1,2 @@
+export * from './Dashboard'
+export * from './Login'
